# backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "EstadoEntidad": {
        "description": "Estado de entidades gestionables (espacios y recursos).",
        "enum": [
          "activo",
          "inactivo",
          "mantenimiento"
        ],
        "title": "EstadoEntidad",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZonaCreate": {
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "espacio_id": {
            "title": "Espacio Id",
            "type": "integer"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad",
            "default": "activo"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Nombre",
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "espacio_id"
        ],
        "title": "ZonaCreate",
        "type": "object"
      },
      "ZonaRecursosResponse": {
        "description": "Respuesta de PUT /zonas/{zona_id}/recursos. Deliberadamente no\nextiende ZonaResponse (que no incluye `recursos` todavía, ver 12C-2) —\nesta subfase agrega su propio schema mínimo en vez de ampliar el\ncontrato ya aprobado de ZonaResponse sin necesidad.",
        "properties": {
          "recurso_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Recurso Ids",
            "type": "array"
          },
          "zona_id": {
            "title": "Zona Id",
            "type": "integer"
          }
        },
        "required": [
          "zona_id",
          "recurso_ids"
        ],
        "title": "ZonaRecursosResponse",
        "type": "object"
      },
      "ZonaRecursosUpdate": {
        "description": "Fase 12C-3: reemplazo completo de la asociación Zona<->Recurso.",
        "properties": {
          "recurso_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Recurso Ids",
            "type": "array"
          }
        },
        "title": "ZonaRecursosUpdate",
        "type": "object"
      },
...
      "ZonaResponse": {
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "integer"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "espacio_id": {
            "title": "Espacio Id",
            "type": "integer"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEntidad"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "recurso_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Recurso Ids",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "nombre",
          "espacio_id",
          "descripcion",
          "capacidad",
          "estado",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by"
        ],
        "title": "ZonaResponse",
        "type": "object"
      },
      "ZonaUpdate": {
        "properties": {
          "capacidad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidad"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "espacio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Espacio Id"
          },
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EstadoEntidad"
              },
              {
                "type": "null"
              }
            ]
          },
          "nombre": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          }
        },
        "title": "ZonaUpdate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API para gestionar recursos y sus reservas",
    "title": "Gestión de Reservas de Recursos Institucionales",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/zonas": {
      "get": {
        "description": "Listar zonas. No requiere autenticacion (mismo criterio RN-005 que\nGET /espacios): anonimo/usuario solo ven zonas activas de espacios\nactivos; gestor/admin ven todo.",
        "operationId": "listar_zonas_zonas_get",
        "parameters": [
          {
            "in": "query",
            "name": "espacio_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Espacio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ZonaResponse"
                  },
                  "title": "Response Listar Zonas Zonas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Zonas",
        "tags": [
          "zonas"
        ]
      },
      "post": {
        "operationId": "crear_zona_zonas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZonaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Crear Zona",
        "tags": [
          "zonas"
        ]
      }
    },
    "/zonas/{zona_id}": {
      "delete": {
        "operationId": "eliminar_zona_zonas__zona_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "zona_id",
            "required": true,
            "schema": {
              "title": "Zona Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Eliminar Zona",
        "tags": [
          "zonas"
        ]
      },
      "put": {
        "operationId": "actualizar_zona_zonas__zona_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "zona_id",
            "required": true,
            "schema": {
              "title": "Zona Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZonaUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Actualizar Zona",
        "tags": [
          "zonas"
        ]
      }
    },
    "/zonas/{zona_id}/recursos": {
      "put": {
        "description": "Reemplazo completo de la asociación Zona<->Recurso (Fase 12C-3):\nla lista recibida sustituye por completo a la actual -- lo ausente se\nelimina, lo nuevo se inserta. Unicidad funcional: un recurso ya\nasociado a OTRA zona responde 409 sin dejar cambios parciales (todas\nlas validaciones se resuelven antes de tocar la base de datos).",
        "operationId": "actualizar_recursos_de_zona_zonas__zona_id__recursos_put",
        "parameters": [
          {
            "in": "path",
            "name": "zona_id",
            "required": true,
            "schema": {
              "title": "Zona Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZonaRecursosUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonaRecursosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Actualizar Recursos De Zona",
        "tags": [
          "zonas"
        ]
      }
    }
  }
}